fix(ci): restore Rust 1.95 warning-clean gate (#11652)
Pin and migrate the complete Homeboy dependency graph reached by homeboy-cli so Rust 1.95 formatting, workspace test compilation, and strict Clippy all pass on current main.

Closes #11580.

AI assistance: OpenAI openai/gpt-5.6-sol and openai/gpt-5.6-terra via OpenCode partitioned the migration into compatibility domains, implemented and reviewed the component changes, reconciled current main, and verified the complete deterministic gate set. Chris Huber remains responsible for every line.

+++ crates/homeboy-core/src/cleanup.rs
@@ -256,7 +256,7 @@ fn filesystem_available_bytes(path: &Path) -> Option<u64> {
     let mut stat = std::mem::MaybeUninit::<libc::statvfs>::uninit();
     (unsafe { libc::statvfs(path.as_ptr(), stat.as_mut_ptr()) } == 0).then(|| unsafe {
         let stat = stat.assume_init();
-        (stat.f_bavail as u64).saturating_mul(stat.f_frsize as u64)
+        (stat.f_bavail as u64).saturating_mul(stat.f_frsize)
     })
 }
